- removed the xid framework and replaced the kiipsdk 3.0.3 version.
