+++ CHANGELOG.md
@@ -1,7 +1,17 @@
-# KiipSDK CHANGELOG
+# CHANGELOG
 
-## 3.0.1
-**Changes:**
+## v3.0.3 - 2018-09-04
+* Added a module map
+* Native View fix
+* Removed XID framework
+* Improved Ad measurability
+* Bug fixes
+
+## v3.0.2 - 2018-06-29
+* Documentation updated to fix app store submission error
+* Bug fixes
+
+## v3.0.1 - 2018-05-03
 * Able to handle different redirect pages
 * Ability to capture more accurate user data
 * Updated Custom Notification in app feature
@@ -10,3 +20,68 @@
 * Integrated the XID SDK
 * Xcode 9.3 or above is required
 * Bug fixes
+
+## v2.3.2 - 2018-01-18
+* Support MOAT Integration
+* Bug fixes
+
+## v2.3.1 - 2017-02-08
+* Streamlined KiipSDK initialization
+* Bug fixes
+
+## v2.3.0 - 2015-09-24
+* Inline video autoplay support
+* Added callback support for video start and end events
+* iOS 9 compatibility and bitcode support
+* New Api 'setTestMode' to receive test rewards during development
+* All communication over secured protocol (https)
+* Moved resource bundle (KiipSDKResources) to KiipSDK.framework level
+* Bug fixes
+
+## v2.2.2 - 2015-05-18
+* Bugfixes
+
+## v2.1.1 - 2014-9-16
+* iOS 8 fixes: landscape, alerts, and views.
+* iOS 8 optimized.
+
+## v2.1.0 - 2014-7-11
+* Remove swarm
+* Remove stock notification .png
+* Bugfixes
+
+## v2.0.8 - 2014-5-16
+* Native Video reward unit
+* Better unique identifier support
+
+## v2.0.7 - 2014-3-31
+* Optimize network calls
+
+## v2.0.6 - 2014-1-29
+* Support 64 bit.
+* Raise min target to 6.0.
+* Bug fixes
+
+## v2.0.5 - 2013-11-26
+* Improved Reward support
+
+## v2.0.4 - 2013-08-19
+* Added Cross Promo
+* Added support for iOS 7
+* Added Passbook integration
+* Optimized network connections
+* Small bug fixes
+
+## v2.0.3 - 2013-05-08
+* Renamed all strings/variables with `uniqueIdentifier`
+
+## v2.0.2 - 2013-02-12
+* Added 'share' capability
+* Small bug fixes
+
+## v2.0.1 - 2012-11-20
+* Added automatic location detection only if the application has already received permissions for location
+* Fix: `saveMoment:value:success:failure` now sends the correct value
+
+## v2.0.0 - 2012-11-16
+* Initial release
