--- Kiip/Framework/KiipSDK.framework/Headers
@@ -1,1 +1,0 @@
-./Versions/A/Headers
