+++ Kiip/Framework/KiipSDK.framework/Headers/KiipSDK.h
@@ -1,0 +1,15 @@
+//
+//  KiipSDK.h
+//  Kiip
+//
+//  Created by Grantland Chew on 10/17/12.
+//  Copyright (c) 2012 Kiip, Inc. All rights reserved.
+//
+
+#import "Kiip.h"
+#import "KPPoptart.h"
+#import "KPNotification.h"
+#import "KPModal.h"
+#import "KPNotificationView.h"
+#import "KPNativeRewardView.h"
+#import "KPTableView.h"
