+++ Kiip/Framework/KiipSDK.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module KiipSDK {
+  umbrella header "KiipSDK.h"
+
+  export *
+  module * { export * }
+}
