+++ Kiip/Framework/KiipSDK.framework/_CodeSignature/CodeResources
@@ -1,0 +1,271 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/KPModal.h</key>
+		<data>
+		MT60hbaDQTvrLl+Udh13ZnjLvwk=
+		</data>
+		<key>Headers/KPNativeRewardView.h</key>
+		<data>
+		Qi8bkQ47bwjw/pu7N0aDKzNlDTc=
+		</data>
+		<key>Headers/KPNotification.h</key>
+		<data>
+		5PmeseBYiH+ENGXxIf3/49ob86w=
+		</data>
+		<key>Headers/KPNotificationView.h</key>
+		<data>
+		2BtSnBfCWjK4oUJNg4k+32FmQEs=
+		</data>
+		<key>Headers/KPPoptart.h</key>
+		<data>
+		U7icvY6CGImxGwA0buuUqd+a2eI=
+		</data>
+		<key>Headers/KPTableView.h</key>
+		<data>
+		grCUQNoTYBWUMKfYsqnqzYjhym4=
+		</data>
+		<key>Headers/KPTableViewCell.h</key>
+		<data>
+		QdeLwYDr6KSy3dyoReSnRwiaFY8=
+		</data>
+		<key>Headers/Kiip.h</key>
+		<data>
+		ZxThAdyW9rBldN7iNjs761qqKoo=
+		</data>
+		<key>Headers/KiipSDK.h</key>
+		<data>
+		a8LiyxZt5TjSmIM8gbwjW9I5GDM=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		9cydS9+dc/0q/Mam/0kEfFr9iho=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		UKCLVCEjUm+Vnj4O/fBj3VIdoIM=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/KPModal.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			MT60hbaDQTvrLl+Udh13ZnjLvwk=
+			</data>
+			<key>hash2</key>
+			<data>
+			GILS/OxSZaySBdjvH08AbpVh9Iq5CKiL0AJnxAbjRBM=
+			</data>
+		</dict>
+		<key>Headers/KPNativeRewardView.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			Qi8bkQ47bwjw/pu7N0aDKzNlDTc=
+			</data>
+			<key>hash2</key>
+			<data>
+			opUGeDEpI51oRpVvjwCgrh/P05zysk9ipPHCsXqqMyU=
+			</data>
+		</dict>
+		<key>Headers/KPNotification.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			5PmeseBYiH+ENGXxIf3/49ob86w=
+			</data>
+			<key>hash2</key>
+			<data>
+			YeVMWGQ5I9RujAzHRtk9RtOmNkE2k0OlGPEnCTXuipk=
+			</data>
+		</dict>
+		<key>Headers/KPNotificationView.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			2BtSnBfCWjK4oUJNg4k+32FmQEs=
+			</data>
+			<key>hash2</key>
+			<data>
+			HA+LQL4mu4FpE5NCYyKglSQEipi1zJn30aZCRG/4tKE=
+			</data>
+		</dict>
+		<key>Headers/KPPoptart.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			U7icvY6CGImxGwA0buuUqd+a2eI=
+			</data>
+			<key>hash2</key>
+			<data>
+			tGW3yEHaZvHD1Zg+loNb7NS4K8ssKLI3/Nh0nmOI0fo=
+			</data>
+		</dict>
+		<key>Headers/KPTableView.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			grCUQNoTYBWUMKfYsqnqzYjhym4=
+			</data>
+			<key>hash2</key>
+			<data>
+			3UpWRRVfts7HFzDeiofy6Wzot0sjSLvx+0wRxyzrfkY=
+			</data>
+		</dict>
+		<key>Headers/KPTableViewCell.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			QdeLwYDr6KSy3dyoReSnRwiaFY8=
+			</data>
+			<key>hash2</key>
+			<data>
+			tH0s+rqp6lu4LwZ8ercoOjcbAg9cTewaPBSZUsKEZBQ=
+			</data>
+		</dict>
+		<key>Headers/Kiip.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			ZxThAdyW9rBldN7iNjs761qqKoo=
+			</data>
+			<key>hash2</key>
+			<data>
+			kkC1UQDsJcKy+9k+G8C/fn8WTzyzhUVoPxm75TLFgZY=
+			</data>
+		</dict>
+		<key>Headers/KiipSDK.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			a8LiyxZt5TjSmIM8gbwjW9I5GDM=
+			</data>
+			<key>hash2</key>
+			<data>
+			fAcse6JbUdR2NqzuojCMr3etYyzfTeLJe2kzy9OK2h8=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			UKCLVCEjUm+Vnj4O/fBj3VIdoIM=
+			</data>
+			<key>hash2</key>
+			<data>
+			SScZrW7Z7305VulhSu8iW9krDvbca5GIfhcIgR/p9rI=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^[^/]+$</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
