--- Kiip/Framework/XIDLibrary.framework/Headers/ObjCObjectFactory.h
@@ -1,9 +1,0 @@
-@import Foundation;
-
-@interface ObjCObjectFactory : NSObject
-
-+ (id)create:(NSString *)className;
-
-+ (id)create:(NSString *)className initializer:(SEL)initializer argument:(id)argument;
-
-@end
