--- Kiip/Framework/XIDLibrary.framework/Headers/XIDLibrary.h
@@ -1,19 +1,0 @@
-//
-//  XIDLibrary.h
-//  XIDLibrary
-//
-//  Created by XID on 2/16/18.
-//  Copyright © 2018 XID. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for XIDLibrary.
-FOUNDATION_EXPORT double XIDLibraryVersionNumber;
-
-//! Project version string for XIDLibrary.
-FOUNDATION_EXPORT const unsigned char XIDLibraryVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <XIDLibrary/PublicHeader.h>
-
-#import <XIDLibrary/ObjCObjectFactory.h>
