--- Kiip/Framework/XIDLibrary.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module XIDLibrary {
-  umbrella header "XIDLibrary.h"
-
-  export *
-  module * { export * }
-}
-
-module XIDLibrary.Swift {
-    header "XIDLibrary-Swift.h"
-    requires objc
-}
