+++ README.md
@@ -1,41 +1,63 @@
-# KiipSDK
+# Kiip iOS SDK
 
-[![Version](http://cocoapod-badges.herokuapp.com/v/KiipSDK/badge.png)](http://cocoadocs.org/docsets/KiipSDK)
-[![Platform](http://cocoapod-badges.herokuapp.com/p/KiipSDK/badge.png)](http://cocoadocs.org/docsets/KiipSDK)
+Supported platforms: iOS 9.3+
+Compiled with support for 64-bit.
 
-## Usage
+## Public interfaces
 
-To run the example project; clone the repo, and run `pod install` from the Example directory first.
+* Kiip.h
+* KPPoptart.h
+* KPNotification.h
+* KPModal.h
+* KPNotificationView.h
+* KPNativeRewardView.h
+* KPTableView.h
 
-## Requirements
+## Resources
 
-Trying to update a pod with teh latest SDK? Ensure Ruby 2.1.1+ is installed, as well as cocoapods:
+* KiipSDKResources.bundle
+* kp_activity_indicator.png
+* kp_activity_indicator@2x.png
+* kp_webview_o_button.png
+* kp_webview_o_button@2x.png
+* kp_webview_x_button.png
+* kp_webview_x_button@2x.png
 
-`rvm update <version>`
-`rvm use <version>`
-`sudo gem install cocoapods`
+## Required Libraries
 
-## Installation
+* CoreTelephony.framework
+* QuartzCore.framework
+* SystemConfiguration.framework
+* AdSupport.framework
+* MediaPlayer.framework
+* AVFoundation.framework
+* CoreMotion.framework
+* GLKit.framework
+* KIIPMoatMobileAppKit.framework
 
-KiipSDK is available through [CocoaPods](http://cocoapods.org), to install
-it simply add the following line to your Podfile:
-    
-    pod "KiipSDK"
+## Optional Libraries
 
-If you have created a Podfile for the first time, uncomment these two lines of code (or add them).
+* CoreLocation.framework
 
-    platform :ios, '9.0'
-    
-    use_frameworks!
+## Instructions
 
-Open up "YOUR_PROJECT_NAME".xcworkspace and click on your project and go to the "Build Setting."
-Under Build Option, there is a field called "Always Embed Swift Standard Libraries" and mark this as "Yes"
+1. Drag-drop `KiipSDK.framework`, `KIIPMoatMobileAppKit.framework`, and `KiipSDKResources.bundle` into your XCode project.
+2. In `Build Phases`, verify that `KiipSDK.framework` and `KIIPMoatMobileAppKit.framework` are in the `Build Phases` > `Link Binary with Libraries` and the `KiipSDKResources.bundle` in `Build Phases` >  `Copy Bundle Resources` section
+3. Add the required libraries to your project (mentioned above)
+4. Use `#import <KiipSDK/KiipSDK.h>`
 
-That's it! You are ready to start your project!
- 
-Visit our [documentation page](http://docs.kiip.me) for more information on how to use it.
+### Location (Optional)
 
-## License
+Kiip has the ability to automatically retrieve the user's location, but the application must first
+request permission for location updates as well as include CoreLocation.framework
 
-KiipSDK is available under the MIT license. See the LICENSE file for more info.
+To request permission, you must add the following code:
 
+CLLocationManager *lm = [[CLLocationManager alloc] init];
+[lm startUpdatingLocation];
+[lm stopUpdatingLocation];
+
+
+## Support
+
+For integration and reference, see our docs at http://docs.kiip.me
